fixed basic routes and api callsgit add -Agit add -A
