--- public/js/keys.js
@@ -1,5 +1,0 @@
-console.log("this is loaded");
-exports.spotify = {
-  id: process.env.SPOTIFY_ID,
-  secret: process.env.SPOTIFY_SECRET,
-};
