+++ routes/keys.js
@@ -1,0 +1,6 @@
+console.log("this is fucking loaded");
+// require("dotenv").config();
+exports.spotify = {
+  id: process.env.SPOTIFY_ID,
+  secret: process.env.SPOTIFY_SECRET
+};
