+++ views/index.handlebars
@@ -40,12 +40,4 @@
             </div>
         </div>
 
-<script>
-  
-  $(document).ready( () => {
-	// on submit..
-	
-  
-
-  });
-</script>
+<script src="./js/index.js"></script>
